requirements.txt
PyQt5>=5.15.0
python-pptx>=0.6.21
Pillow>=9.0.0
lxml>=4.6.0
